requirements.txt
biopython>=1.83,<2.0
numpy>=1.26.4,<2.0
matplotlib>=3.8.0,<4.0
pyyaml>=6.0.1
requests>=2.31.0
